LayoutTests/media/video-source-error-no-candidate.html

<html lang="en">
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
        <title>&lt;video&gt; and &lt;source&gt; error test</title>
        <script src=video-test.js></script>

        <script>

            var sources = [];
            var errorCount = 0;

            function errorEvent(evt)
            {
                consoleWrite("");

                var ndx;
                for (ndx = 0; ndx < sources.length; ndx++) {
                    if (sources[ndx] == evt.target)
                        break;
                }

                if (sources[ndx] == evt.target)
                    logResult(true, "EVENT(error) from &lt;source id='<em>" + evt.target.id + "</em>' src='<em>" + relativeURL(evt.target.src) + "</em>'&gt;");
                else
                    logResult(false, "EVENT(error) from " + evt.target);

                testExpected("video.error", null);
                
                if (++errorCount == 3) {
                    consoleWrite("");
                    setTimeout(endTest, 1000);
                }
            }

            function start()
            {
                document.addEventListener("error", errorEvent, true);

                sources = document.getElementsByTagName('source');

                findMediaElement();
                waitForEvent("loadstart");
            }
        </script>
    </head>

    <body>

       <video onloadstart="start()" controls>
            <source id=missing-src type="video/blahblah">
            <source id=bogus-type src=content/test.mp4 type="video/blahblah">
            <source id=unsupported-media-query src=content/test.mp4 media=print>
        </video>
        
        <p>Test that 'error' events are fired when no urls are passed to the media engine.</p>

    </body>
</html>
